requirements.txt
pandas
numpy
requests
astronomer-cosmos
